requirements.txt
streamlit==1.51.0
python-dotenv==1.2.1
pdfplumber==0.11.7
scikit-learn==1.7.2
sentence-transformers==5.1.2
werkzeug==3.1.3
langchain==1.0.5
langchain_community==0.4.1
langchain-groq==1.0.0
faiss-cpu==1.12.0
python-docx==1.2.0